fix(check_cell_distribution):require cell division
The check_cell_distribution unit test now runs the matcha main
function and verifies whether the cells collection has been
divvied up across images.  The check sums the number of cells
across images and checks whether it matches the total.  In the
spirit of test-driven development, the test now fails and thus
provides a specification for the feature to add next:
cell-collection distribution in the matcha function.