refac(t_cell_collection):def history in subroutine
This version uses a subroutine to define history elements in order
to eliminate unnecessary object temporaries and copies in the
matcha function for performance reasons.  The change provided
no single-image performance improvement. We'll tag this commit
to preserve it for future reference and further testing, e.g.,
in multi-image execution.